Move all widget related components into their own folder
