+++ src/components/profile/index.ts
@@ -1,30 +1,30 @@
 import { Profile } from "./profile";
 import { ProfileHeader } from "./profile-header"; 
-import { Widget } from "./widget";
-import { WidgetContainer } from "./widget-container"; 
-import { ResizeWidget } from "./resize-widget"; 
-import { AddWidgets } from "./add-widgets"; 
+import { Widget } from "./widgets/widget";
+import { WidgetContainer } from "./widgets/widget-container"; 
+import { ResizeWidget } from "./widgets/resize-widget"; 
+import { AddWidgets } from "./widgets/add-widgets"; 
 import { InvalidAlert } from "./invalid-alert"; 
 import { ProfileEditModal } from "./profile-edit-modal";
 import { InputLocation } from "./input-location"; 
 import { InputSkills } from "./input-skills"; 
-import { YouTubeIframe } from "./youtube-iframe"; 
-import { WidgetHeader } from "./widget-header"; 
-import { DefaultWidget } from "./widget-default"; 
-import { DribbbleWidget } from "./widget-dribbble"; 
-import { BehanceWidget } from "./widget-behance"; 
-import { MediumWidget } from "./widget-medium"; 
-import { YouTubeWidget } from "./widget-youtube"; 
-import { SubstackWidget } from "./widget-substack"; 
-import { SpotifyAlbumWidget } from "./widget-spotify-album"; 
-import { SpotifySongWidget } from "./widget-spotify-song"; 
-import { SoundcloudWidget } from "./widget-soundcloud"; 
-import { InstagramWidget } from "./widget-instagram"; 
-import { GithubWidget } from "./widget-github"; 
-import { FigmaWidget } from "./widget-figma"; 
-import { TwitterWidget } from "./widget-twitter"; 
-import { PhotoWidget } from "./widget-photo"; 
-import { NoWidgetsContent } from "./no-widgets-content"; 
+import { YouTubeIframe } from "./widgets/youtube-iframe"; 
+import { WidgetHeader } from "./widgets/widget-header"; 
+import { DefaultWidget } from "./widgets/widget-default"; 
+import { DribbbleWidget } from "./widgets/widget-dribbble"; 
+import { BehanceWidget } from "./widgets/widget-behance"; 
+import { MediumWidget } from "./widgets/widget-medium"; 
+import { YouTubeWidget } from "./widgets/widget-youtube"; 
+import { SubstackWidget } from "./widgets/widget-substack"; 
+import { SpotifyAlbumWidget } from "./widgets/widget-spotify-album"; 
+import { SpotifySongWidget } from "./widgets/widget-spotify-song"; 
+import { SoundcloudWidget } from "./widgets/widget-soundcloud"; 
+import { InstagramWidget } from "./widgets/widget-instagram"; 
+import { GithubWidget } from "./widgets/widget-github"; 
+import { FigmaWidget } from "./widgets/widget-figma"; 
+import { TwitterWidget } from "./widgets/widget-twitter"; 
+import { PhotoWidget } from "./widgets/widget-photo"; 
+import { NoWidgetsContent } from "./widgets/no-widgets-content"; 
 
 export { 
   Profile,
