+++ src/components/profile/widgets/resize-widget.tsx
