+++ src/components/profile/widgets/widget-default.tsx
@@ -1,4 +1,4 @@
-import { WidgetHeader } from '@/components/profile/widget-header';
+import { WidgetHeader } from '@/components/profile/widgets/widget-header';
 import { WidgetType } from '@/types';
 import React from 'react';
