+++ src/components/profile/widgets/widget-figma.tsx
@@ -1,36 +1,32 @@
 import { ImageOverlay } from '@/components/common';
-import { WidgetHeader } from '@/components/profile/widget-header';
-import { MediumArticleContentType, WidgetSize, WidgetType } from '@/types';
-import React, { useEffect } from 'react';
+import { WidgetHeader } from '@/components/profile/widgets/widget-header';
+import { FigmaWidgetContentType, WidgetSize, WidgetType } from '@/types';
+import React from 'react';
 
-interface MediumWidgetType {
-  content: MediumArticleContentType;
+interface FigmaWidgetType {
+  content: FigmaWidgetContentType;
   size: WidgetSize;
 }
 
-export const MediumWidget: React.FC<MediumWidgetType> = ({ content, size }) => {
-  useEffect(() => {
-    console.log(size.toString());
-  }, [size]);
-
+export const FigmaWidget: React.FC<FigmaWidgetType> = ({ content, size }) => {
   let widgetLayout;
   switch (size) {
     case WidgetSize.A:
       widgetLayout = (
         <div className='h-full flex flex-col gap-2'>
           <div className='flex flex-row gap-2'>
-            <div className='w-1/4'>
-              <WidgetHeader type={WidgetType.Medium} />
+            <div className='w-10'>
+              <WidgetHeader type={WidgetType.Figma} />
             </div>
             <div>
               <div className='text-sm font-semibold'>{content.title}</div>
-              <div className='text-xs text-gray-500'>{content.host}</div>
+              <div className='text-xs text-gray-500'>{content.description}</div>
             </div>
           </div>
           <div className='flex-grow relative rounded-xl overflow-hidden'>
             <img
               src={content.image}
-              alt='Medium content'
+              alt='Figma content'
               className='absolute inset-0 w-full h-full object-cover'
             />
             <ImageOverlay />
@@ -42,16 +38,16 @@ export const MediumWidget: React.FC<MediumWidgetType> = ({ content, size }) => {
       widgetLayout = (
         <div className='h-full flex flex-col gap-2'>
           <div>
-            <WidgetHeader type={WidgetType.Medium} noMargin />
+            <WidgetHeader type={WidgetType.Figma} noMargin />
           </div>
           <div>
             <div className='text-sm font-semibold'>{content.title}</div>
-            <div className='text-xs text-gray-500'>{content.host}</div>
+            <div className='text-xs text-gray-500'>{content.description}</div>
           </div>
           <div className='h-full w-full relative rounded-xl overflow-hidden'>
             <img
               src={content.image}
-              alt='Medium content'
+              alt='Figma content'
               className='w-full h-full object-cover'
             />
             <ImageOverlay />
@@ -63,14 +59,14 @@ export const MediumWidget: React.FC<MediumWidgetType> = ({ content, size }) => {
       widgetLayout = (
         <div className='h-full flex flex-row gap-2'>
           <div className='w-2/5'>
-            <WidgetHeader type={WidgetType.Medium} />
+            <WidgetHeader type={WidgetType.Figma} />
             <div className='text-sm font-semibold'>{content.title}</div>
-            <div className='text-xs text-gray-500'>{content.host}</div>
+            <div className='text-xs text-gray-500'>{content.description}</div>
           </div>
           <div className={`relative rounded-xl overflow-hidden w-3/5`}>
             <img
               src={content.image}
-              alt='Medium content'
+              alt='Figma content'
               className='w-full h-full object-cover'
               style={{ objectFit: 'cover' }}
             />
@@ -82,15 +78,15 @@ export const MediumWidget: React.FC<MediumWidgetType> = ({ content, size }) => {
     case WidgetSize.D:
       widgetLayout = (
         <div className='h-full flex flex-col gap-2'>
-          <WidgetHeader type={WidgetType.Medium} noMargin />
+          <WidgetHeader type={WidgetType.Figma} noMargin />
           <div>
             <div className='text-sm font-semibold'>{content.title}</div>
-            <div className='text-xs text-gray-500'>{content.host}</div>
+            <div className='text-xs text-gray-500'>{content.description}</div>
           </div>
           <div className={`h-full w-full relative rounded-xl overflow-hidden`}>
             <img
               src={content.image}
-              alt='Medium content'
+              alt='Figma content'
               className='w-full h-full object-cover'
               style={{ objectFit: 'cover' }}
             />
