+++ src/components/profile/widgets/widget-github.tsx
@@ -1,41 +1,32 @@
 import { ImageOverlay } from '@/components/common';
-import { WidgetHeader } from '@/components/profile/widget-header';
-import { SubstackWidgetContentType, WidgetSize, WidgetType } from '@/types';
+import { WidgetHeader } from '@/components/profile/widgets/widget-header';
+import { GithubWidgetContentType, WidgetSize, WidgetType } from '@/types';
 import React, { useEffect } from 'react';
 
-interface SubstackWidgetType {
-  content: SubstackWidgetContentType;
+interface GithubWidgetType {
+  content: GithubWidgetContentType;
   size: WidgetSize;
 }
 
-export const SubstackWidget: React.FC<SubstackWidgetType> = ({
-  content,
-  size,
-}) => {
+export const GithubWidget: React.FC<GithubWidgetType> = ({ content, size }) => {
   let widgetLayout;
-  const currentType = WidgetType.Substack;
-
-  const localHeader = (
-    <>
-      <div className='text-sm font-semibold'>{content.title}</div>
-      <div className='text-xs text-gray-500'>{content.host}</div>
-    </>
-  );
-
   switch (size) {
     case WidgetSize.A:
       widgetLayout = (
         <div className='h-full flex flex-col gap-2'>
           <div className='flex flex-row gap-2'>
-            <div className='w-8'>
-              <WidgetHeader type={currentType} />
+            <div className='w-10'>
+              <WidgetHeader type={WidgetType.Github} />
             </div>
-            <div className='w-90'>{localHeader}</div>
+            <div>
+              <div className='text-sm font-semibold'>{content.title}</div>
+              <div className='text-xs text-gray-500'>github.com</div>
+            </div>
           </div>
           <div className='flex-grow relative rounded-xl overflow-hidden'>
             <img
               src={content.image}
-              alt='Substack content'
+              alt='Github content'
               className='absolute inset-0 w-full h-full object-cover'
             />
             <ImageOverlay />
@@ -47,13 +38,16 @@ export const SubstackWidget: React.FC<SubstackWidgetType> = ({
       widgetLayout = (
         <div className='h-full flex flex-col gap-2'>
           <div>
-            <WidgetHeader type={currentType} noMargin />
+            <WidgetHeader type={WidgetType.Github} noMargin />
           </div>
-          <div>{localHeader}</div>
+          <div>
+            <div className='text-sm font-semibold'>{content.title}</div>
+            <div className='text-xs text-gray-500'>github.com</div>
+          </div>
           <div className='h-full w-full relative rounded-xl overflow-hidden'>
             <img
               src={content.image}
-              alt='Substack content'
+              alt='Github content'
               className='w-full h-full object-cover'
             />
             <ImageOverlay />
@@ -65,13 +59,14 @@ export const SubstackWidget: React.FC<SubstackWidgetType> = ({
       widgetLayout = (
         <div className='h-full flex flex-row gap-2'>
           <div className='w-2/5'>
-            <WidgetHeader type={currentType} />
-            {localHeader}
+            <WidgetHeader type={WidgetType.Github} />
+            <div className='text-sm font-semibold'>{content.title}</div>
+            <div className='text-xs text-gray-500'>github.com</div>
           </div>
           <div className={`relative rounded-xl overflow-hidden w-3/5`}>
             <img
               src={content.image}
-              alt='Substack content'
+              alt='Github content'
               className='w-full h-full object-cover'
               style={{ objectFit: 'cover' }}
             />
@@ -83,12 +78,15 @@ export const SubstackWidget: React.FC<SubstackWidgetType> = ({
     case WidgetSize.D:
       widgetLayout = (
         <div className='h-full flex flex-col gap-2'>
-          <WidgetHeader type={currentType} noMargin />
-          <div>{localHeader}</div>
+          <WidgetHeader type={WidgetType.Github} noMargin />
+          <div>
+            <div className='text-sm font-semibold'>{content.title}</div>
+            <div className='text-xs text-gray-500'>github.com</div>
+          </div>
           <div className={`h-full w-full relative rounded-xl overflow-hidden`}>
             <img
               src={content.image}
-              alt='Substack content'
+              alt='Github content'
               className='w-full h-full object-cover'
               style={{ objectFit: 'cover' }}
             />
