+++ src/components/profile/widgets/widget-medium.tsx
@@ -1,44 +1,35 @@
 import { ImageOverlay } from '@/components/common';
-import { WidgetHeader } from '@/components/profile/widget-header';
-import { WidgetSize, WidgetType, YoutubeWidgetContentType } from '@/types';
+import { WidgetHeader } from '@/components/profile/widgets/widget-header';
+import { MediumArticleContentType, WidgetSize, WidgetType } from '@/types';
 import React, { useEffect } from 'react';
 
-interface YouTubeWidgetType {
-  content: YoutubeWidgetContentType;
+interface MediumWidgetType {
+  content: MediumArticleContentType;
   size: WidgetSize;
 }
 
-export const YouTubeWidget: React.FC<YouTubeWidgetType> = ({
-  content,
-  size,
-}) => {
+export const MediumWidget: React.FC<MediumWidgetType> = ({ content, size }) => {
   useEffect(() => {
     console.log(size.toString());
   }, [size]);
 
   let widgetLayout;
-  const currentType = WidgetType.Youtube;
-
-  const localHeader = (
-    <>
-      <div className='text-sm font-semibold'>{content.title}</div>
-      <div className='text-xs text-gray-500'>www.youtube.com</div>
-    </>
-  );
-
   switch (size) {
     case WidgetSize.A:
       widgetLayout = (
         <div className='h-full flex flex-col gap-2'>
           <div className='flex flex-row gap-2'>
             <div className='w-1/4'>
-              <WidgetHeader type={currentType} />
+              <WidgetHeader type={WidgetType.Medium} />
             </div>
-            <div>{localHeader}</div>
+            <div>
+              <div className='text-sm font-semibold'>{content.title}</div>
+              <div className='text-xs text-gray-500'>{content.host}</div>
+            </div>
           </div>
           <div className='flex-grow relative rounded-xl overflow-hidden'>
             <img
-              src={content.thumbnail}
+              src={content.image}
               alt='Medium content'
               className='absolute inset-0 w-full h-full object-cover'
             />
@@ -51,12 +42,15 @@ export const YouTubeWidget: React.FC<YouTubeWidgetType> = ({
       widgetLayout = (
         <div className='h-full flex flex-col gap-2'>
           <div>
-            <WidgetHeader type={currentType} noMargin />
+            <WidgetHeader type={WidgetType.Medium} noMargin />
           </div>
-          <div>{localHeader}</div>
+          <div>
+            <div className='text-sm font-semibold'>{content.title}</div>
+            <div className='text-xs text-gray-500'>{content.host}</div>
+          </div>
           <div className='h-full w-full relative rounded-xl overflow-hidden'>
             <img
-              src={content.thumbnail}
+              src={content.image}
               alt='Medium content'
               className='w-full h-full object-cover'
             />
@@ -69,12 +63,13 @@ export const YouTubeWidget: React.FC<YouTubeWidgetType> = ({
       widgetLayout = (
         <div className='h-full flex flex-row gap-2'>
           <div className='w-2/5'>
-            <WidgetHeader type={currentType} />
-            {localHeader}
+            <WidgetHeader type={WidgetType.Medium} />
+            <div className='text-sm font-semibold'>{content.title}</div>
+            <div className='text-xs text-gray-500'>{content.host}</div>
           </div>
           <div className={`relative rounded-xl overflow-hidden w-3/5`}>
             <img
-              src={content.thumbnail}
+              src={content.image}
               alt='Medium content'
               className='w-full h-full object-cover'
               style={{ objectFit: 'cover' }}
@@ -87,11 +82,14 @@ export const YouTubeWidget: React.FC<YouTubeWidgetType> = ({
     case WidgetSize.D:
       widgetLayout = (
         <div className='h-full flex flex-col gap-2'>
-          <WidgetHeader type={currentType} noMargin />
-          <div>{localHeader}</div>
+          <WidgetHeader type={WidgetType.Medium} noMargin />
+          <div>
+            <div className='text-sm font-semibold'>{content.title}</div>
+            <div className='text-xs text-gray-500'>{content.host}</div>
+          </div>
           <div className={`h-full w-full relative rounded-xl overflow-hidden`}>
             <img
-              src={content.thumbnail}
+              src={content.image}
               alt='Medium content'
               className='w-full h-full object-cover'
               style={{ objectFit: 'cover' }}
