+++ src/components/profile/widgets/widget-photo.tsx
