+++ src/components/profile/widgets/widget-spotify-album.tsx
@@ -1,5 +1,5 @@
 import { ImageOverlay } from '@/components/common';
-import { WidgetHeader } from '@/components/profile/widget-header';
+import { WidgetHeader } from '@/components/profile/widgets/widget-header';
 import { SpotifyWidgetContentType, WidgetSize, WidgetType } from '@/types';
 import { Play } from 'lucide-react';
 import React, { useEffect } from 'react';
@@ -9,7 +9,7 @@ interface SpotifyWidgetType {
   size: WidgetSize;
 }
 
-export const SpotifySongWidget: React.FC<SpotifyWidgetType> = ({
+export const SpotifyAlbumWidget: React.FC<SpotifyWidgetType> = ({
   content,
   size,
 }) => {
@@ -18,7 +18,7 @@ export const SpotifySongWidget: React.FC<SpotifyWidgetType> = ({
   }, [size]);
 
   let widgetLayout;
-  const currentType = WidgetType.SpotifySong;
+  const currentType = WidgetType.SpotifyAlbum;
 
   const localHeader = (
     <>
@@ -55,17 +55,16 @@ export const SpotifySongWidget: React.FC<SpotifyWidgetType> = ({
         <div className='h-full flex flex-col gap-2'>
           <div className='flex justify-between'>
             <WidgetHeader type={currentType} noMargin />
-            <button className='cursor-pointer flex gap-1 items-center bg-[#573DF5] text-white px-4 text-sm py-1 rounded-lg'>
-              <Play size={16} />
-              Play
-            </button>
           </div>
           <div>{localHeader}</div>
           <div className='h-full w-full relative rounded-xl overflow-hidden mt-6 bg-white text-black'>
-            <img
-              src={content.cover}
-              alt='Spotify content'
-              className='absolute inset-0 w-full h-full object-cover'
+            <iframe
+              title='Spotify'
+              className='SpotifyPlayer absolute w-full h-full'
+              src={`https://embed.spotify.com/?uri=${content.url}&view=list&theme=light`}
+              width='100%'
+              height='450px'
+              style={{ border: 'none', background: 'white' }}
             />
             <ImageOverlay />
           </div>
