+++ src/components/profile/widgets/widget-youtube.tsx
@@ -1,32 +1,45 @@
 import { ImageOverlay } from '@/components/common';
-import { WidgetHeader } from '@/components/profile/widget-header';
-import { FigmaWidgetContentType, WidgetSize, WidgetType } from '@/types';
-import React from 'react';
+import { WidgetHeader } from '@/components/profile/widgets/widget-header';
+import { WidgetSize, WidgetType, YoutubeWidgetContentType } from '@/types';
+import React, { useEffect } from 'react';
 
-interface FigmaWidgetType {
-  content: FigmaWidgetContentType;
+interface YouTubeWidgetType {
+  content: YoutubeWidgetContentType;
   size: WidgetSize;
 }
 
-export const FigmaWidget: React.FC<FigmaWidgetType> = ({ content, size }) => {
+export const YouTubeWidget: React.FC<YouTubeWidgetType> = ({
+  content,
+  size,
+}) => {
+  useEffect(() => {
+    console.log(size.toString());
+  }, [size]);
+
   let widgetLayout;
+  const currentType = WidgetType.Youtube;
+
+  const localHeader = (
+    <>
+      <div className='text-sm font-semibold'>{content.title}</div>
+      <div className='text-xs text-gray-500'>www.youtube.com</div>
+    </>
+  );
+
   switch (size) {
     case WidgetSize.A:
       widgetLayout = (
         <div className='h-full flex flex-col gap-2'>
           <div className='flex flex-row gap-2'>
-            <div className='w-10'>
-              <WidgetHeader type={WidgetType.Figma} />
+            <div className='w-1/4'>
+              <WidgetHeader type={currentType} />
             </div>
-            <div>
-              <div className='text-sm font-semibold'>{content.title}</div>
-              <div className='text-xs text-gray-500'>{content.description}</div>
-            </div>
+            <div>{localHeader}</div>
           </div>
           <div className='flex-grow relative rounded-xl overflow-hidden'>
             <img
-              src={content.image}
-              alt='Figma content'
+              src={content.thumbnail}
+              alt='Medium content'
               className='absolute inset-0 w-full h-full object-cover'
             />
             <ImageOverlay />
@@ -38,16 +51,13 @@ export const FigmaWidget: React.FC<FigmaWidgetType> = ({ content, size }) => {
       widgetLayout = (
         <div className='h-full flex flex-col gap-2'>
           <div>
-            <WidgetHeader type={WidgetType.Figma} noMargin />
+            <WidgetHeader type={currentType} noMargin />
           </div>
-          <div>
-            <div className='text-sm font-semibold'>{content.title}</div>
-            <div className='text-xs text-gray-500'>{content.description}</div>
-          </div>
+          <div>{localHeader}</div>
           <div className='h-full w-full relative rounded-xl overflow-hidden'>
             <img
-              src={content.image}
-              alt='Figma content'
+              src={content.thumbnail}
+              alt='Medium content'
               className='w-full h-full object-cover'
             />
             <ImageOverlay />
@@ -59,14 +69,13 @@ export const FigmaWidget: React.FC<FigmaWidgetType> = ({ content, size }) => {
       widgetLayout = (
         <div className='h-full flex flex-row gap-2'>
           <div className='w-2/5'>
-            <WidgetHeader type={WidgetType.Figma} />
-            <div className='text-sm font-semibold'>{content.title}</div>
-            <div className='text-xs text-gray-500'>{content.description}</div>
+            <WidgetHeader type={currentType} />
+            {localHeader}
           </div>
           <div className={`relative rounded-xl overflow-hidden w-3/5`}>
             <img
-              src={content.image}
-              alt='Figma content'
+              src={content.thumbnail}
+              alt='Medium content'
               className='w-full h-full object-cover'
               style={{ objectFit: 'cover' }}
             />
@@ -78,15 +87,12 @@ export const FigmaWidget: React.FC<FigmaWidgetType> = ({ content, size }) => {
     case WidgetSize.D:
       widgetLayout = (
         <div className='h-full flex flex-col gap-2'>
-          <WidgetHeader type={WidgetType.Figma} noMargin />
-          <div>
-            <div className='text-sm font-semibold'>{content.title}</div>
-            <div className='text-xs text-gray-500'>{content.description}</div>
-          </div>
+          <WidgetHeader type={currentType} noMargin />
+          <div>{localHeader}</div>
           <div className={`h-full w-full relative rounded-xl overflow-hidden`}>
             <img
-              src={content.image}
-              alt='Figma content'
+              src={content.thumbnail}
+              alt='Medium content'
               className='w-full h-full object-cover'
               style={{ objectFit: 'cover' }}
             />
